fix: parse power levels as js_int integers, accepting strings
Room versions before 10 allow power levels as strings, and all room versions
allow negative ones, neither of which parsed as u64. Power levels copied into
the new room are rewritten as integers, which room version 10 and later
require.

+++ src/config.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 
+use js_int::Int;
 use serde::Deserialize;
 
 #[derive(Deserialize)]
@@ -8,7 +9,7 @@ pub struct Config {
     pub access_token: String,
     pub target_room_version: usize,
     pub rooms: Vec<String>,
-    pub pl_overrides: HashMap<String, u64>,
+    pub pl_overrides: HashMap<String, Int>,
     pub state_events_to_transfer: Vec<String>,
     pub drop_members: Vec<String>,
 }
